restructured site
some css switched to sass, media queries next to come as well as tasks
outlined for me yesterday. I'm fooling around with the links too, I
think it would be best if we uploaded the site to my server so that I
can correct the links more easily, sound good? text me if you would
prefer to stick with yours. will complete remainder over this aft.